# refactor: DataElement merge, use SQL when merging DataValues and EventDataValues [DHIS2-14429]