--- a/dhis-2/dhis-api/src/main/java/org/hisp/dhis/datavalue/DataValueStore.java
+++ b/dhis-2/dhis-api/src/main/java/org/hisp/dhis/datavalue/DataValueStore.java
@@ -260,4 +260,27 @@ public interface DataValueStore {
    */
   void mergeDataValuesWithAttributeOptionCombos(
       @Nonnull CategoryOptionCombo target, @Nonnull Collection<CategoryOptionCombo> sources);
+
+  /**
+   * SQL for handling merging {@link DataValue}s. There may be multiple potential {@link DataValue}
+   * duplicates. Duplicate {@link DataValue}s with the latest {@link DataValue#lastUpdated} values
+   * are kept, the rest are deleted. Only one of these entries can exist due to the composite key
+   * constraint. <br>
+   * The 3 execution paths are:
+   *
+   * <p>1. If the source {@link DataValue} is not a duplicate, it simply gets its {@link
+   * DataValue#dataElement} updated to that of the target.
+   *
+   * <p>2. If the source {@link DataValue} is a duplicate and has an earlier {@link
+   * DataValue#lastUpdated} value, it is deleted.
+   *
+   * <p>3. If the source {@link DataValue} is a duplicate and has a later {@link
+   * DataValue#lastUpdated} value, the target {@link DataValue} is deleted. The source is kept and
+   * has its {@link DataValue#dataElement} updated to that of the target.
+   *
+   * @param target target {@link DataElement}
+   * @param sources source {@link DataElement}s
+   */
+  void mergeDataValuesWithDataElements(
+      @Nonnull DataElement target, @Nonnull Collection<DataElement> sources);
 }