--- a/dhis-2/dhis-services/dhis-service-administration/src/main/java/org/hisp/dhis/merge/dataelement/handler/DataDataElementMergeHandler.java
+++ b/dhis-2/dhis-services/dhis-service-administration/src/main/java/org/hisp/dhis/merge/dataelement/handler/DataDataElementMergeHandler.java
@@ -75,7 +75,7 @@ public class DataDataElementMergeHandler {
               + " dataMergeStrategy being used, deleting source data values");
       dataValueStore.deleteDataValues(sources);
     } else {
-      // TODO dv store merge DEs
+      dataValueStore.mergeDataValuesWithDataElements(target, sources);
     }
   }